refactor(discovery): decompose the fingerprint god-file by role (WS-D)
Split internal/discovery/fingerprint.go (786 LOC) into role-based files in
the same package — pure move: core probing/port-scan orchestration stays
in fingerprint.go; OS fingerprinting, service-version detection, and TLS
probing move to their own files.

All files under the 600-line size-gate threshold. Declaration sets and
code-line multisets verified identical to the original.

Related to #187
