+++ internal/discovery/fingerprint.go
@@ -16,10 +16,8 @@ package discovery
 
 import (
 	"context"
-	"crypto/tls"
 	"fmt"
 	"net"
-	"regexp"
 	"slices"
 	"strings"
 	"sync"
@@ -212,367 +210,6 @@ func (f *Fingerprinter) fingerprintOS(
 	}
 
 	return fp
-}
-
-// fingerprintFromBanners analyzes port banners for OS information.
-func (f *Fingerprinter) fingerprintFromBanners(profile *DeviceProfile, fp *OSFingerprint) {
-	for _, port := range profile.OpenPorts {
-		if port.Banner == "" {
-			continue
-		}
-		bannerLower := strings.ToLower(port.Banner)
-		if osInfo := f.parseOSFromBanner(bannerLower); osInfo != nil {
-			f.updateFingerprintIfBetter(fp, osInfo, "banner")
-		}
-	}
-}
-
-// fingerprintFromHTTP analyzes HTTP Server header for OS information.
-func (f *Fingerprinter) fingerprintFromHTTP(profile *DeviceProfile, fp *OSFingerprint) {
-	if profile.HTTPInfo == nil || profile.HTTPInfo.Server == "" {
-		return
-	}
-	serverLower := strings.ToLower(profile.HTTPInfo.Server)
-	if osInfo := f.parseOSFromServer(serverLower); osInfo != nil {
-		f.updateFingerprintIfBetter(fp, osInfo, "http")
-	}
-}
-
-// updateFingerprintIfBetter updates the fingerprint if the new info has higher confidence.
-func (*Fingerprinter) updateFingerprintIfBetter(fp, osInfo *OSFingerprint, method string) {
-	fp.Methods = append(fp.Methods, method)
-	if osInfo.Confidence > fp.Confidence {
-		fp.OSFamily = osInfo.OSFamily
-		fp.OSVersion = osInfo.OSVersion
-		fp.Confidence = osInfo.Confidence
-	}
-}
-
-// osMatch defines a pattern for OS detection.
-type osMatch struct {
-	patterns   []string // all patterns must match
-	osFamily   string
-	osVersion  string
-	confidence int
-}
-
-// getSSHOSMatchers returns OS patterns for SSH banners.
-func getSSHOSMatchers() []osMatch {
-	return []osMatch{
-		{[]string{"ubuntu"}, osLinux, "ubuntu", 90},
-		{[]string{"debian"}, osLinux, "debian", 90},
-		{[]string{"centos"}, osLinux, "rhel", 90},
-		{[]string{"red hat"}, osLinux, "rhel", 90},
-		{[]string{"freebsd"}, "bsd", "freebsd", 90},
-		{[]string{"cisco"}, osCisco, "", 95},
-		{[]string{"windows"}, osWindows, "", 90},
-		{[]string{"openssh"}, "unix", "", 50},
-	}
-}
-
-// getGenericOSMatchers returns OS patterns for generic banners.
-func getGenericOSMatchers() []osMatch {
-	return []osMatch{
-		{[]string{"linux"}, osLinux, "", 80},
-		{[]string{"windows"}, osWindows, "", 80},
-		{[]string{"cisco"}, osCisco, "", 95},
-		{[]string{"junos"}, "juniper", "", 95},
-		{[]string{"vsftpd"}, osLinux, "", 75},
-		{[]string{"proftpd"}, osLinux, "", 75},
-		{[]string{"microsoft", "ftp"}, osWindows, "", 85},
-	}
-}
-
-// parseOSFromBanner extracts OS info from service banners.
-func (f *Fingerprinter) parseOSFromBanner(banner string) *OSFingerprint {
-	fp := &OSFingerprint{}
-
-	if strings.Contains(banner, "ssh") {
-		f.matchOSPatterns(banner, getSSHOSMatchers(), fp)
-	}
-	if fp.OSFamily == "" {
-		f.matchOSPatterns(banner, getGenericOSMatchers(), fp)
-	}
-
-	if fp.OSFamily == "" {
-		return nil
-	}
-	return fp
-}
-
-// matchOSPatterns checks banner against patterns and sets fingerprint if matched.
-func (*Fingerprinter) matchOSPatterns(banner string, matchers []osMatch, fp *OSFingerprint) {
-	for _, m := range matchers {
-		matched := true
-		for _, p := range m.patterns {
-			if !strings.Contains(banner, p) {
-				matched = false
-				break
-			}
-		}
-		if matched {
-			fp.OSFamily = m.osFamily
-			fp.OSVersion = m.osVersion
-			fp.Confidence = m.confidence
-			return
-		}
-	}
-}
-
-// getServerOSMatchers returns OS patterns for HTTP Server headers.
-func getServerOSMatchers() []osMatch {
-	return []osMatch{
-		{[]string{"ubuntu"}, osLinux, "ubuntu", 85},
-		{[]string{"debian"}, osLinux, "debian", 85},
-		{[]string{"centos"}, osLinux, "rhel", 85},
-		{[]string{"red hat"}, osLinux, "rhel", 85},
-		{[]string{"cisco"}, osCisco, "", 90},
-		{
-			[]string{"routeros"},
-			"mikrotik",
-			"",
-			95,
-		}, //nolint:misspell // RouterOS is MikroTik's product name
-		{[]string{"fortinet"}, "fortinet", "", 95},
-		{[]string{"fortigate"}, "fortinet", "", 95},
-		{[]string{"pfsense"}, "bsd", "firewall", 90},
-		{[]string{"opnsense"}, "bsd", "firewall", 90},
-		{[]string{"synology"}, osLinux, "dsm", 95},
-		{[]string{"qnap"}, osLinux, "qts", 95},
-	}
-}
-
-// parseOSFromServer extracts OS info from HTTP Server header.
-func (f *Fingerprinter) parseOSFromServer(server string) *OSFingerprint {
-	fp := &OSFingerprint{}
-
-	// Windows indicators (special case for IIS version extraction)
-	if strings.Contains(server, "microsoft") || strings.Contains(server, "iis") {
-		fp.OSFamily = osWindows
-		if match := regexp.MustCompile(`iis[/\s]*([\d.]+)`).FindStringSubmatch(server); len(
-			match,
-		) > 1 {
-			fp.OSVersion = "IIS " + match[1]
-		}
-		fp.Confidence = 85
-		return fp
-	}
-
-	// Try pattern matching
-	f.matchOSPatterns(server, getServerOSMatchers(), fp)
-
-	// Fallback for generic web servers
-	if fp.OSFamily == "" &&
-		(strings.Contains(server, "lighttpd") || strings.Contains(server, "nginx")) {
-		fp.OSFamily = "unix"
-		fp.Confidence = 50
-	}
-
-	if fp.OSFamily == "" {
-		return nil
-	}
-	return fp
-}
-
-// detectServiceVersion analyzes a port's banner to determine service version.
-func (f *Fingerprinter) detectServiceVersion(port OpenPort) *ServiceVersion {
-	if port.Banner == "" && port.Service == "" {
-		return nil
-	}
-
-	sv := &ServiceVersion{
-		Port:       port.Port,
-		Service:    port.Service,
-		Confidence: defaultServiceConfidence,
-	}
-
-	if port.Banner == "" {
-		return sv
-	}
-
-	bannerLower := strings.ToLower(port.Banner)
-	f.detectSSHVersion(port.Port, bannerLower, sv)
-	f.detectFTPVersion(port.Port, bannerLower, sv)
-	f.detectSMTPVersion(port.Port, bannerLower, sv)
-	f.detectTelnetVersion(port.Port, bannerLower, sv)
-
-	return sv
-}
-
-// detectSSHVersion detects SSH service version from banner.
-func (*Fingerprinter) detectSSHVersion(port int, banner string, sv *ServiceVersion) {
-	if port != 22 && !strings.Contains(banner, "ssh") {
-		return
-	}
-	sv.Service = "ssh"
-	if match := regexp.MustCompile(`openssh[_\s]*([\d.p]+)`).FindStringSubmatch(banner); len(
-		match,
-	) > 1 {
-		sv.Product = "OpenSSH"
-		sv.Version = match[1]
-		sv.Confidence = 95
-	} else if sshMatch := regexp.MustCompile(`ssh-([\d.]+)`).FindStringSubmatch(banner); len(sshMatch) > 1 {
-		sv.Product = "SSH"
-		sv.Version = sshMatch[1]
-		sv.Confidence = 80
-	}
-}
-
-// detectFTPVersion detects FTP service version from banner.
-func (*Fingerprinter) detectFTPVersion(port int, banner string, sv *ServiceVersion) {
-	if port != 21 && !strings.HasPrefix(banner, "220") {
-		return
-	}
-	sv.Service = "ftp"
-	switch {
-	case strings.Contains(banner, "vsftpd"):
-		sv.Product = "vsftpd"
-		if match := regexp.MustCompile(`vsftpd\s*([\d.]+)`).FindStringSubmatch(banner); len(
-			match,
-		) > 1 {
-			sv.Version = match[1]
-		}
-		sv.Confidence = 90
-	case strings.Contains(banner, "proftpd"):
-		sv.Product = "ProFTPD"
-		if match := regexp.MustCompile(`proftpd\s*([\d.]+)`).FindStringSubmatch(banner); len(
-			match,
-		) > 1 {
-			sv.Version = match[1]
-		}
-		sv.Confidence = 90
-	case strings.Contains(banner, "pure-ftpd"):
-		sv.Product = "Pure-FTPd"
-		sv.Confidence = 90
-	case strings.Contains(banner, "microsoft"):
-		sv.Product = "Microsoft FTP"
-		sv.Confidence = 85
-	}
-}
-
-// detectSMTPVersion detects SMTP service version from banner.
-func (*Fingerprinter) detectSMTPVersion(port int, banner string, sv *ServiceVersion) {
-	if port != 25 && port != 587 {
-		return
-	}
-	sv.Service = "smtp"
-	switch {
-	case strings.Contains(banner, "postfix"):
-		sv.Product = "Postfix"
-		sv.Confidence = 90
-	case strings.Contains(banner, "sendmail"):
-		sv.Product = "Sendmail"
-		sv.Confidence = 90
-	case strings.Contains(banner, "exim"):
-		sv.Product = "Exim"
-		if match := regexp.MustCompile(`exim\s*([\d.]+)`).FindStringSubmatch(banner); len(
-			match,
-		) > 1 {
-			sv.Version = match[1]
-		}
-		sv.Confidence = 90
-	case strings.Contains(banner, "microsoft"):
-		sv.Product = "Microsoft Exchange"
-		sv.Confidence = 85
-	}
-}
-
-// detectTelnetVersion detects Telnet service version from banner.
-func (*Fingerprinter) detectTelnetVersion(port int, banner string, sv *ServiceVersion) {
-	if port != telnetPort {
-		return
-	}
-	sv.Service = "telnet"
-	switch {
-	case strings.Contains(banner, "cisco"):
-		sv.Product = productCiscoIOS
-		sv.Confidence = 95
-	case strings.Contains(banner, "linux"):
-		sv.Product = "Linux telnetd"
-		sv.Confidence = 80
-	}
-}
-
-// probeTLS probes a port for TLS certificate information.
-func (f *Fingerprinter) probeTLS(ctx context.Context, ip string, port int) *TLSInfo {
-	addr := fmt.Sprintf("%s:%d", ip, port)
-
-	// Use context-aware dialing
-	dialer := &net.Dialer{Timeout: f.timeout}
-	rawConn, err := dialer.DialContext(ctx, "tcp", addr)
-	if err != nil {
-		return nil
-	}
-	// Wrap the raw connection with TLS
-	tlsConfig := &tls.Config{
-		InsecureSkipVerify: true, // #nosec G402 -- We want to inspect any certificate
-	}
-	conn := tls.Client(rawConn, tlsConfig)
-	if handshakeErr := conn.HandshakeContext(ctx); handshakeErr != nil {
-		_ = rawConn.Close()
-		return nil
-	}
-	defer func() { _ = conn.Close() }()
-
-	state := conn.ConnectionState()
-	if len(state.PeerCertificates) == 0 {
-		return nil
-	}
-
-	cert := state.PeerCertificates[0]
-
-	info := &TLSInfo{
-		Port:              port,
-		Version:           tlsVersionString(state.Version),
-		CipherSuite:       tls.CipherSuiteName(state.CipherSuite),
-		CommonName:        cert.Subject.CommonName,
-		ValidFrom:         cert.NotBefore,
-		ValidTo:           cert.NotAfter,
-		DaysUntilExpiry:   int(time.Until(cert.NotAfter).Hours() / hoursPerDay),
-		SelfSigned:        cert.Issuer.CommonName == cert.Subject.CommonName,
-		SubjectAltNames:   cert.DNSNames,
-		CertificateErrors: []string{},
-	}
-
-	// Build issuer string
-	if len(cert.Issuer.Organization) > 0 {
-		info.Issuer = cert.Issuer.Organization[0]
-	} else {
-		info.Issuer = cert.Issuer.CommonName
-	}
-
-	// Check for common certificate issues
-	now := time.Now()
-	if now.Before(cert.NotBefore) {
-		info.CertificateErrors = append(info.CertificateErrors, "not yet valid")
-	}
-	if now.After(cert.NotAfter) {
-		info.CertificateErrors = append(info.CertificateErrors, "expired")
-	}
-	if info.SelfSigned {
-		info.CertificateErrors = append(info.CertificateErrors, "self-signed")
-	}
-	if info.DaysUntilExpiry > 0 && info.DaysUntilExpiry < 30 {
-		info.CertificateErrors = append(info.CertificateErrors, "expiring soon")
-	}
-
-	return info
-}
-
-// tlsVersionString converts TLS version constant to string.
-func tlsVersionString(version uint16) string {
-	switch version {
-	case tls.VersionTLS10:
-		return "TLS 1.0"
-	case tls.VersionTLS11:
-		return "TLS 1.1"
-	case tls.VersionTLS12:
-		return "TLS 1.2"
-	case tls.VersionTLS13:
-		return "TLS 1.3"
-	default:
-		return fmt.Sprintf("Unknown (0x%04x)", version)
-	}
 }
 
 // containsInt checks if an int slice contains a value.
